decouple modal from main.js with its own module
