+++ public/main.js
@@ -1,3 +1,5 @@
+import { closeModal, openModal } from "./modal.js";
+
 const message = document.getElementById("message-div");
 const phoneNumberDisplay = document.getElementById("phone-number");
 const usernameDisplay = document.getElementById("username");
@@ -143,37 +145,7 @@ const createModifySubscriber = async () => {
   clearDisplay();
 };
 
-const openModal = async (phoneNumberInput) => {
-  document.getElementById("modal-overlay").style.display = "block";
-  document.getElementById("modal-number-display").textContent =
-    phoneNumberInput;
-
-  // Fill in modal input fields
-  const response = await fetch(`/ims/subscriber/${phoneNumberInput}`);
-  const responseJSON = await response.json();
-
-  if (response.status === 200) {
-    document.getElementById("input-username").value = responseJSON.username;
-    document.getElementById("input-password").value = "******";
-    document.getElementById("input-domain").value = responseJSON.domain;
-    document.getElementById("input-status").value = responseJSON.status;
-  } else {
-    message.textContent = "";
-  }
-};
-
-const closeModal = () => {
-  document.getElementById("modal-overlay").style.display = "none";
-
-  // Reset modal input fields to blank on close
-  document.getElementById("modal-number-display").textContent = "";
-  document.getElementById("input-username").value = "";
-  document.getElementById("input-password").value = "";
-  document.getElementById("input-domain").value = "";
-  document.getElementById("input-status").value = "";
-};
-
-clearDisplay = () => {
+const clearDisplay = () => {
   phoneNumberDisplay.textContent = "";
   usernameDisplay.textContent = "";
   passwordDisplay.textContent = "";
